make maintainer-clean before testing and quote variables

+++ cron/nightly/libvirt_cron_testing.sh
@@ -25,11 +25,11 @@ source_dist()
     autogen_f=$srcdir/$autogen
     echo "execute $autogen_f" | tee -a $logfile
 
-    if [ ! -e $autogen_f ] || [ ! -x $autogen_f ]; then
+    if [ ! -e "$autogen_f" ] || [ ! -x "$autogen_f" ]; then
         echo "autogen.sh in $srdir doesn't exist or couldn't be executed" | tee -a $logfile
         return 1
     else
-        (cd $distdir
+        (cd "$distdir"
         $autogen_f
         if [ $? -ne 0 ]; then
             exit 1
@@ -52,13 +52,13 @@ git_pull()
     srcdir=$1
     shift
 
-    if [ ! -d $srcdir/.git ]; then
+    if [ ! -d "$srcdir/.git" ]; then
         echo "$srcdir is not a git direcotry" | tee -a $logfile
         return 1
     else
         echo "cd $srcdir" | tee -a $logfile
         echo "running \"git pull\"..." | tee -a $logfile
-        (cd $srcdir
+        (cd "$srcdir"
          git pull) | tee -a $logfile
 
         if [  $? -ne 0 ]; then
@@ -79,14 +79,14 @@ dist_check()
 distribute()
 {
     rm -rf $targetdir/libvirt-*.tar.gz
-    if ! cp -f $distdir/libvirt-*.tar.gz $targetdir > /dev/null; then
+    if ! cp -f $distdir/libvirt-*.tar.gz "$targetdir" > /dev/null; then
         return 1
     fi
 }
 
 machine_check()
 {
-    local m=$1
+    local m="$1"
     shift
 
     if ! ping -c 3 $m > /dev/null; then
@@ -97,7 +97,7 @@ machine_check()
 autotest_prepare()
 {
     rm_auto_cmd="rm -rf $autotestdir"
-    $SSH -l root $MACHINE $rm_auto_cmd | tee -a $logfile
+    "$SSH" -l root "$MACHINE" "$rm_auto_cmd" | tee -a $logfile
     if [ $? -ne 0 ]; then
         return 1
     fi
@@ -106,19 +106,19 @@ autotest_prepare()
 submit_job()
 {
     retval=0
-    cli=$autotestcli/atest
-    START_DATE=`date +%m/%d/%Y" "%T`
+    cli="$autotestcli/atest"
+    START_DATE=$(date +%m/%d/%Y" "%T)
     echo "Job start at $START_DATE" | tee -a $logfile
 
-    for testsuit in $(echo $TESTSUITS | tr "," "\n")
+    for testsuit in $(echo "$TESTSUITS" | tr "," "\n")
     do
         echo "Running: $PYTHON $cli job create -t $testsuit \
 -m $MACHINE \"$TESTNAME($testsuit)\" --web=$AUTOTEST_SERVER" | tee -a $logfile
-        ($PYTHON $cli job create \
-                             -t $testsuit \
-                             -m $MACHINE \
+        ("$PYTHON" "$cli" job create \
+                             -t "$testsuit" \
+                             -m "$MACHINE" \
                              "$TESTNAME($testsuit)" \
-                             --web=$AUTOTEST_SERVER
+                             --web="$AUTOTEST_SERVER"
         ) | tee -a $logfile
         if [ $? -ne 0 ]; then
             echo "$testsuit submission Failed" | tee -a $logfile
@@ -131,28 +131,28 @@ submit_job()
         sleep_time=1
         while [ 1 ]
         do
-            running_job=$($PYTHON $cli job list -r -u autotest --web=$AUTOTEST_SERVER | sed -n '$p')
+            running_job=$("$PYTHON" "$cli" job list -r -u autotest --web="$AUTOTEST_SERVER" | sed -n '$p')
             if [ -z "$running_job" ]; then
-                result_state=$($PYTHON $cli job list -u autotest --web=$AUTOTEST_SERVER | sed -n '$p')
+                result_state=$("$PYTHON" "$cli" job list -u autotest --web="$AUTOTEST_SERVER" | sed -n '$p')
                 echo "$testsuit :$result_state" | tee -a $logfile
 
-                if echo $result_state | grep "Completed" > /dev/null; then
+                if echo "$result_state" | grep "Completed" > /dev/null; then
                     # sleep 10 secs for testing machine boot
                     echo "Waiting 1 min for testing machine reboot"
                     sleep 60
-                    if $SSH -l root $MACHINE ls $install_flag > /dev/null; then
+                    if "$SSH" -l root "$MACHINE" ls "$install_flag" > /dev/null; then
                         break
                     else
                         echo "Libvirt install testing FAIL!" | tee -a $logfile
                         retval=1
                         break 2
                     fi
-                elif echo $result_state | grep "Failed" > /dev/null; then
+                elif echo "$result_state" | grep "Failed" > /dev/null; then
                     retval=1
                     break 2
                 fi
             else
-                if [ $sleep_time -eq 1 ] && echo $running_job | grep "Running" > /dev/null; then
+                if [ "$sleep_time" -eq 1 ] && echo "$running_job" | grep "Running" > /dev/null; then
                     echo "$testsuit :$running_job" | tee -a $logfile
                     sleep_time=10
                 fi
@@ -161,7 +161,7 @@ submit_job()
         done
     done
 
-    END_DATE=`date +%m/%d/%Y" "%T`
+    END_DATE=$(date +%m/%d/%Y" "%T)
     echo "Job end at $END_DATE" | tee -a $logfile
     echo "$retval"
     return $retval
@@ -169,10 +169,21 @@ submit_job()
 
 init()
 {
-   :>$logfile
+   : >$logfile
    echo "Initiate testing environment" | tee -a $logfile
-   echo "remove old $distdir" | tee -a $logfile
-   rm -rf $distdir
+
+   if [ -d $distdir ]; then
+       if [ -f  $distdir/Makefile ]; then
+           echo "execute \"make maintainer-clean...\"" | tee -a $logfile
+           current_d=$(pwd)
+           cd "$distdir"
+           make maintainer-clean > /dev/null;
+           cd "$current_d"
+       fi
+
+       echo "remove old $distdir" | tee -a $logfile
+       rm -rf $distdir
+   fi
    echo "make new $distdir" | tee -a $logfile
    mkdir $distdir
 }
@@ -202,13 +213,13 @@ do
         exit 1
         ;;
     d)
-        srcdir=$OPTARG
+        srcdir="$OPTARG"
         ;;
     m)
-        MACHINE=$OPTARG
+        MACHINE="$OPTARG"
         ;;
     t)
-        TESTSUITS=$OPTARG
+        TESTSUITS="$OPTARG"
         ;;
     *)
         echo "Invaid option"
@@ -218,19 +229,19 @@ do
     esac
 done
 
-if [ -z $MACHINE ]; then
+if [ -z "$MACHINE" ]; then
     echo "A host is required to run job"
     exit 1
 fi
 echo "The testing machine is $MACHINE"
 
-if [ -z $TESTSUITS ]; then
+if [ -z "$TESTSUITS" ]; then
     echo "A testsuit must be specified"
     exit 1
 fi
 echo "Testsuit is $TESTSUITS"
 
-if ! machine_check $MACHINE; then
+if ! machine_check "$MACHINE"; then
     echo "Failed to ping machine $MACHINE" | tee -a $logfile
     exit 1
 fi
@@ -241,7 +252,7 @@ srcdir=${srcdir-$defaultsrcdir}
 echo "Libvirt source code direcotry is $srcdir" | tee -a $logfile
 echo "" >> $logfile
 
-git_pull $srcdir
+git_pull "$srcdir"
 if [ $? -ne 0 ]; then
     echo "Failed to pull git update" | tee -a $logfile
     exit 1
@@ -250,7 +261,7 @@ echo "Success to git pull" | tee -a $logfile
 echo "" >> $logfile
 
 echo "Make distribution..." | tee -a $logfile
-source_dist $srcdir
+source_dist "$srcdir"
 if [ $? -ne 0 ]; then
     echo "Failed to make distribution" | tee -a $logfile
     exit 1
